Initial support for multiple function returns
The types.Function datatype now has a list of return types instead of a
single return type. This updates it to follow the current parser
behavior and the future direction for the language.

Multiple returns are not yet implemented though. An attempt to declare a
function with multiple return types will crash the compiler with a "not
implemented" type error.