Fix merge conflict
This one was because there was some confusion about what the `tag`
variable meant. In the main branch it was used here to contain the Lua
name for the type tag while in the rest of the code we tend to use this
variable name to refer to the _tag field in our data types.