c-parser: improve construction of type objects
This is the ugliest part of the c-parser code by far, and probably the
buggiest as well. It is full of type() checks because the c99 parser produces
lots of nested tables, and cpp flattens the tree structure aggressively,
causing many one-element arrays to disappear. Properly typed objects
would go a long way in making this code cleaner.