Implement a rehash function for table reads
I put this inside ltable.c because it relies on a lot of ltable.c internal
functions. If in the future we conclude that we really want to use this
function (instead of something a bit different) then it will probably be a good
idea to rethink where to put this in.