c-parser cleanups: fix expression trees
This is the first step in the cleanup of the data structures
produced by c-parser.

The overall goal of this cleanup is to eventually remove special cases in
`cpp`, `cdriver` and `ctypes`, and produce better data structures for use in
`titan-compiler.foreigntypes`.

This step removes the `cpp.remove_wrapping_subtables` function, which was too
eager and converted all singleton sequences `{x}` into `x`, meaning that lots
of places that used these structures needed to check for types of attributes
using `type()`.

At this point, expression nodes have been cleaned up (with the help of the
`typed` module). Type nodes are next.