support for compiling stand-alone programs
This adds support for compiling Titan programs as stand-alone executables.

If a module contains a `main` function with signature `function({string}):integer`,
that is the main entry point of the program. The input argument is an array
of strings containing the command-line arguments, and the integer return value
is the exit code of the program.

Only one of the modules given to `titanc` may contain a `main` function.

Additionally, this commits adds a `-v`/`--verbose` flag to `titanc`, which
makes it display the C compiler invocations.