resolve references to imports when compiled as a single executable
When compiling a program as a single executable, do not use dynamic
linking to resolve references in imports. Instead, generate
simple 'extern' declarations that are resolved in the linking step.

Modify the `examples.sdl_demo` example to demonstrate this,
moving the helper function into a separate module.

Also, add documentation on stand-alone compilation and the
`main` function.